exporters/channel-health-exporter/requirements.txt
prometheus-client==0.19.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
